+++ R/Swrappers.R
@@ -78,7 +78,10 @@ NULL
 #'   value(s) must be of length one. See corresponding section for further
 #'   information.
 #' @param \dots Further arguments passed on to `fun`.
-#' @param cols A character vector specifying the columns to aggregate.
+#' @param cols A character vector specifying the columns to aggregate. Another
+#'   possibility is a character string containing either comma separated column
+#'   names, for example, `"x,y,z"`, or the start and end column separated by a
+#'   colon, for example, `"x:z"`.
 #' @param n A logical specifying if a column named `.n` giving the number of
 #'   values per temporal aggregation level shall be added. See corresponding
 #'   section for further information.
@@ -348,10 +351,15 @@ colapply <- function(x, ...) {
 #'
 #' @param fun A [`function`]. Its return value must be of length one.
 #' @param cols A character vector specifying the columns to apply `fun` to.
+#'   Another possibility is a character string containing either comma separated
+#'   column names, for example, `"x,y,z"`, or the start and end column separated
+#'   by a colon, for example, `"x:z"`.
 #' @param resultCols An optional character vector of the same length as `cols`
-#'   specifying the column names for the return values of `fun`. Non-existing
-#'   columns are added and existing columns are overwritten. Columns are matched
-#'   element-wise between `cols` and `resultCols`.
+#'   specifying the column names for the return values of `fun`. Another
+#'   possibility is a character string containing comma separated column names,
+#'   for example, `"x,y,z"`. Non-existing columns are added and existing columns
+#'   are overwritten. Columns are matched element-wise between `cols` and
+#'   `resultCols`.
 #' @param suffix An optional character string. The return values of `fun` are
 #'   added as new columns with names consisting of the columns specified in
 #'   `cols` and this suffix. Existing columns are never overwritten. Only used
@@ -524,7 +532,9 @@ nas <- function(x, ...) {
 #' recognised periodicity.
 #'
 #' @param cols A character vector specifying the columns whose missing values
-#'   shall be listed.
+#'   shall be listed. Another possibility is a character string containing
+#'   either comma separated column names, for example, `"x,y,z"`, or the start
+#'   and end column separated by a colon, for example, `"x:z"`.
 #' @inheritParams alter.DTSg
 #'
 #' @return Returns a [`data.table::data.table`] with five columns:
@@ -594,9 +604,14 @@ setMethod(
 #'   a character string coercible to one. The data is plotted up to this
 #'   timestamp.
 #' @param cols A character vector specifying the columns whose values shall be
-#'   plotted.
+#'   plotted. Another possibility is a character string containing either comma
+#'   separated column names, for example, `"x,y,z"`, or the start and end column
+#'   separated by a colon, for example, `"x:z"`.
 #' @param secAxisCols An optional character vector specifying the columns whose
-#'   values shall be plotted on a secondary axis. Must be a subset of `cols`.
+#'   values shall be plotted on a secondary axis. Another possibility is a
+#'   character string containing either comma separated column names, for
+#'   example, `"x,y,z"`, or the start and end column separated by a colon, for
+#'   example, `"x:z"`. Must be a subset of `cols`.
 #' @param secAxisLabel A character string specifying the label of the optional
 #'   secondary axis.
 #' @inheritParams alter.DTSg
@@ -691,7 +706,9 @@ rollapply <- function(x, ...) {
 #' [`DTSg`] object with recognised periodicity.
 #'
 #' @param cols A character vector specifying the columns whose rolling window
-#'   `fun` shall be applied to.
+#'   `fun` shall be applied to. Another possibility is a character string
+#'   containing either comma separated column names, for example, `"x,y,z"`, or
+#'   the start and end column separated by a colon, for example, `"x:z"`.
 #' @param before An integerish value specifying the size of the window in time
 #'   steps before the \dQuote{center} of the rolling window.
 #' @param after An integerish value specifying the size of the window in time
@@ -777,6 +794,9 @@ rowaggregate <- function(x, ...) {
 #'   all the values of the specified `cols`. The return value(s) must be of
 #'   length one. See corresponding section for further information.
 #' @param cols A character vector specifying the columns to apply `fun` to.
+#'   Another possibility is a character string containing either comma separated
+#'   column names, for example, `"x,y,z"`, or the start and end column separated
+#'   by a colon, for example, `"x:z"`.
 #' @inheritParams aggregate.DTSg
 #'
 #' @section Summary functions:
@@ -869,9 +889,13 @@ setColNames <- function(x, ...) {
 #' Changes the column names of [`DTSg`] objects.
 #'
 #' @param cols A character vector specifying the columns whose names shall be
-#'   set. The name of the _.dateTime_ column cannot be changed.
+#'   set. Another possibility is a character string containing either comma
+#'   separated column names, for example, `"x,y,z"`, or the start and end column
+#'   separated by a colon, for example, `"x:z"`. The name of the _.dateTime_
+#'   column cannot be changed.
 #' @param values A character vector of the same length as `cols` specifying the
-#'   desired column names.
+#'   desired column names. Another possibility is a character string containing
+#'   comma separated column names, for example, `"x,y,z"`.
 #' @inheritParams alter.DTSg
 #'
 #' @inherit alter.DTSg return
@@ -916,7 +940,9 @@ setCols <- function(x, ...) {
 #'   argument of [`data.table::data.table`]. Filter expressions can contain the
 #'   special symbol [`.N`][data.table::special-symbols].
 #' @param cols A character vector specifying the columns whose values shall be
-#'   set. The values of the _.dateTime_ column cannot be changed.
+#'   set. Another possibility is a character string containing comma separated
+#'   column names, for example, `"x,y,z"`. The values of the _.dateTime_ column
+#'   cannot be changed.
 #' @param values A vector, [`list`] or list-like object (e.g.
 #'   [`data.table::data.table`]) of replacement and/or new values accepted by
 #'   the `value` argument of \pkg{data.table}'s [`data.table::set`] function.
@@ -973,8 +999,11 @@ setCols.DTSg <- S3WrapperGenerator(expression(DTSg$public_methods$setCols))
 #'
 #' Filters rows and/or selects columns of a [`DTSg`] object.
 #'
-#' @param cols A character vector specifying the columns to select. The
-#'   _.dateTime_ column is always selected and cannot be part of it.
+#' @param cols A character vector specifying the columns to select. Another
+#'   possibility is a character string containing either comma separated column
+#'   names, for example, `"x,y,z"`, or the start and end column separated by a
+#'   colon, for example, `"x:z"`. The _.dateTime_ column is always selected and
+#'   cannot be part of it.
 #' @param funby One of the temporal aggregation level functions described in
 #'   [`TALFs`] or a user defined temporal aggregation level function. Can be
 #'   used to, for instance, select the last two observations of a certain
@@ -1039,7 +1068,9 @@ subset.DTSg <- S3WrapperGenerator(expression(DTSg$public_methods$subset))
 #'
 #' @param object A [`DTSg`] object (S3 method only).
 #' @param cols A character vector specifying the columns whose values shall be
-#'   summarised.
+#'   summarised. Another possibility is a character string containing either
+#'   comma separated column names, for example, `"x,y,z"`, or the start and end
+#'   column separated by a colon, for example, `"x:z"`.
 #' @param \dots Further arguments passed on to [`summary.data.frame`].
 #'
 #' @return Returns a [`table`].